Add HybridSharedSecret type and move hybrid types to utils module
- Create HybridSharedSecret type for consistency with other hybrid types
- Move HybridEncapsulationKey, HybridDecapsulationKey, HybridCiphertext from ghp to utils
- Update all hybrid KEM implementations to use HybridSharedSecret instead of Vec<u8>
- Update PRE to import types from utils instead of ghp module
- Export common hybrid types from lib.rs for better API organization
- All tests pass with improved type consistency
