+++ reference-implementation/src/lib.rs
@@ -47,3 +47,4 @@ mod hybrid_tests;
 // Re-export commonly used items
 pub use error::KemError;
 pub use traits::{EncapsDerand, HybridKemLabel, Kdf, Kem, NominalGroup, Prg};
+pub use utils::{HybridCiphertext, HybridDecapsulationKey, HybridEncapsulationKey, HybridSharedSecret};
